+++ src/arg_parse.c
@@ -99,7 +99,7 @@ char *parse_args(int argc, char **argv, arp_cmd_args_t *out_args) {
                 } else if (CMP_LONG_FLAG(flag, flag_len, FLAG_OUTPUT_LONG)) {
                     out_args->output_path = param;
                 } else if (CMP_LONG_FLAG(flag, flag_len, FLAG_PART_SIZE_LONG)) {
-                    size_t param_l = strtoull(param, NULL, BASE_10);
+                    uint64_t param_l = strtoull(param, NULL, BASE_10);
 
                     if (errno != 0) {
                         return _parse_failed("Invalid param '%s' for flag '%s'", param, arg);
@@ -140,7 +140,7 @@ char *parse_args(int argc, char **argv, arp_cmd_args_t *out_args) {
                 } else if (strcmp(flag, FLAG_OUTPUT_SHORT) == 0) {
                     out_args->output_path = param;
                 } else if (strcmp(flag, FLAG_PART_SIZE_SHORT) == 0) {
-                    size_t param_l = strtoull(param, NULL, BASE_10);
+                    uint64_t param_l = strtoull(param, NULL, BASE_10);
 
                     if (errno != 0) {
                         return _parse_failed("Invalid param '%s' for flag '%s'", param, arg);
